# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pullRequest/create": {
            "post": {
                "description": "Создаёт новый Pull Request от указанного автора.\nРевьюеры выбираются автоматически на основе команды автора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Создать Pull Request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "токен администратора(Вводить без Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для создания Pull Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePR"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PR успешно создан",
                        "schema": {
                            "$ref": "#/definitions/dto.PullRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Автор или команда не найдены",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "PR с таким ID уже существует",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера"
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "description": "Пометить PR как MERGED",
                "tags": [
                    "PullRequests"
                ],
                "summary": "Пометить PR как MERGED",
                "parameters": [
                    {
                        "description": "Id PR",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MergeRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "токен пользователя(Вводить без Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PR в состоянии MERGED",
                        "schema": {
                            "$ref": "#/definitions/dto.MergeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "description": "Переназначить конкретного ревьювера на другого из его команды",
                "tags": [
                    "PullRequests"
                ],
                "summary": "Переназначить конкретного ревьювера на другого из его команды",
                "parameters": [
                    {
                        "description": "data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReassignPullRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "токен администратора(Вводить без Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Переназначение выполнено",
                        "schema": {
                            "$ref": "#/definitions/dto.MergeResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "PR не найден",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/add": {
            "post": {
                "description": "Создаёт новую команду. Если пользователь уже в другой команде, PR пользователя переназначается на участников старой команды.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "Создание новой команды",
                "parameters": [
                    {
                        "description": "Данные команды",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Команда создана",
                        "schema": {
                            "$ref": "#/definitions/dto.TeamResponse"
                        }
                    },
                    "400": {
                        "description": "Команда уже существует или некоректный запрос",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/get_team/{team_name}": {
            "get": {
                "description": "Возвращает информацию о команде и её участниках. Доступ разрешён только для участников команды.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "Получение информации о команде",
                "parameters": [
                    {
                        "type": "string",
                        "description": "токен пользователя(Вводить без Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Уникальное имя команды",
                        "name": "team_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обьект команды",
                        "schema": {
                            "$ref": "#/definitions/dto.TeamResponse"
                        }
                    },
                    "400": {
                        "description": "Пустое имя команды",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён, пользователь не в команде",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Команда не найдена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddTeamRequest": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MemberDto"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.CreatePR": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorMessage"
                }
            }
        },
...
        "dto.MergeRequest": {
            "type": "object",
            "properties": {
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "dto.MergeResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/dto.MergedPullRequestOfUser"
                }
            }
        },
...
        "dto.PullRequest": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.PullRequestResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/dto.PullRequest"
                }
            }
        },
        "dto.ReassignPullRequest": {
            "type": "object",
            "properties": {
                "old_reviewer_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.TeamResponse": {
            "type": "object",
            "properties": {
                "team": {
                    "$ref": "#/definitions/dto.TeamDtoResponse"
                }
            }
        },
...
    }
}